refactor(dashboard): move Auto-Dream runtime panel from Settings to Memory page
The Auto-Dream section is a runtime control surface (per-agent enrollment,
trigger / abort live dreams, status badges) — not a configuration knob. It
fit awkwardly next to theme/language/TOTP/Config-Backup on SettingsPage.
MemoryPage is the natural home: auto-dream is memory consolidation, and
the page already aggregates memory stats and per-agent KV state.

i18n keys remain under settings.auto_dream_* to keep the locale diff to
zero; the user-visible strings are unchanged.
