+++ crates/librefang-api/dashboard/src/pages/SettingsPage.tsx
@@ -5,22 +5,15 @@ import { Button } from "../components/ui/Button";
 import { Badge } from "../components/ui/Badge";
 import {
   Globe, Sun, Moon, Settings, PanelLeftClose, PanelLeft, Languages, LayoutDashboard,
-  Shield, CheckCircle, XCircle, Download, Play, Square, Eye, EyeOff,
+  Shield, CheckCircle, XCircle, Download, Eye, EyeOff,
 } from "lucide-react";
 import { useUIStore } from "../lib/store";
-import { useAutoDreamStatus } from "../lib/queries/autoDream";
 import { useTotpStatus } from "../lib/queries/approvals";
-import {
-  useTriggerAutoDream,
-  useAbortAutoDream,
-  useSetAutoDreamEnabled,
-} from "../lib/mutations/autoDream";
 import {
   useTotpSetup,
   useTotpConfirm,
   useTotpRevoke,
 } from "../lib/mutations/approvals";
-import type { AutoDreamAgentStatus } from "../api";
 
 interface SegmentOption<T extends string> {
   value: T;
@@ -172,9 +165,6 @@ export function SettingsPage() {
 
       {/* Config Backup */}
       <ConfigBackupSection />
-
-      {/* Auto-Dream (background memory consolidation) */}
-      <AutoDreamSection />
     </div>
   );
 }
@@ -522,428 +512,6 @@ function ConfigBackupSection() {
             {t("settings.export_config_btn", "Download")}
           </a>
         </SettingRow>
-      </div>
-    </div>
-  );
-}
-
-/* ------------------------------------------------------------------ */
-/*  Auto-Dream Section                                                 */
-/* ------------------------------------------------------------------ */
-
-// Format an epoch-ms into a short human-readable "N hours ago" / "in N
-// hours" label. Returns "never" when ts is 0 or undefined — the status
-// endpoint omits `next_eligible_at_ms` for never-dreamed agents, and
-// `last_consolidated_at_ms` is 0 in the same case.
-const _rtfCache = new Map<string, Intl.RelativeTimeFormat>();
-
-function getRelativeTimeFormat(locale: string): Intl.RelativeTimeFormat {
-  let rtf = _rtfCache.get(locale);
-  if (!rtf) {
-    rtf = new Intl.RelativeTimeFormat(locale, {
-      numeric: "auto",
-      style: "narrow",
-    });
-    _rtfCache.set(locale, rtf);
-  }
-  return rtf;
-}
-
-function formatRelativeMs(
-  ts: number | undefined,
-  now: number,
-  locale: string,
-  t: (key: string) => string,
-): string {
-  if (ts === undefined || ts === 0) return t("common.never");
-  const diff = ts - now;
-  const absMinutes = Math.abs(diff) / 60_000;
-  const rtf = getRelativeTimeFormat(locale);
-  if (absMinutes < 60) {
-    return rtf.format(Math.round(diff / 60_000), "minute");
-  }
-  const absHours = absMinutes / 60;
-  if (absHours < 24) {
-    return rtf.format(parseFloat((diff / 3_600_000).toFixed(1)), "hour");
-  }
-  return rtf.format(parseFloat((diff / 86_400_000).toFixed(1)), "day");
-}
-
-// Human-readable duration for effective_min_hours. Switches between hours,
-// days, and weeks so "every 168h" renders as "every 1w" etc.
-function formatHours(hours: number, t: (key: string) => string): string {
-  if (hours < 1) return `${(hours * 60).toFixed(0)}${t("settings.auto_dream_dur_minute")}`;
-  if (hours < 24) return `${hours % 1 === 0 ? hours.toFixed(0) : hours.toFixed(1)}${t("settings.auto_dream_dur_hour")}`;
-  const days = hours / 24;
-  if (days < 7) return `${days % 1 === 0 ? days.toFixed(0) : days.toFixed(1)}${t("settings.auto_dream_dur_day")}`;
-  const weeks = days / 7;
-  return `${weeks % 1 === 0 ? weeks.toFixed(0) : weeks.toFixed(1)}${t("settings.auto_dream_dur_week")}`;
-}
-
-function AutoDreamAgentRow({
-  agent,
-  disabled,
-  onTrigger,
-  onAbort,
-  onToggle,
-  triggerPending,
-  abortPending,
-  togglePending,
-}: {
-  agent: AutoDreamAgentStatus;
-  disabled: boolean;
-  onTrigger: (id: string) => void;
-  onAbort: (id: string) => void;
-  onToggle: (id: string, enabled: boolean) => void;
-  triggerPending: boolean;
-  abortPending: boolean;
-  togglePending: boolean;
-}) {
-  const { t, i18n } = useTranslation();
-  const now = Date.now();
-  const progress = agent.progress;
-  const running = progress?.status === "running";
-  const lastTurn = progress?.turns[progress.turns.length - 1];
-  const optedIn = agent.auto_dream_enabled;
-
-  return (
-    <div className="rounded-lg border border-border-subtle/50 bg-main">
-      <div className="flex items-center justify-between px-3 py-2">
-        <div className="flex items-start gap-2 min-w-0 flex-1">
-          <Moon
-            className={`w-4 h-4 shrink-0 mt-0.5 ${
-              optedIn
-                ? running
-                  ? "text-purple-400 animate-pulse"
-                  : "text-purple-400"
-                : "text-text-dim"
-            }`}
-          />
-          <div className="min-w-0 flex-1">
-            <div className="flex items-center gap-2">
-              <p className="text-sm font-medium truncate">{agent.agent_name}</p>
-              {progress && (
-                <Badge
-                  variant={
-                    progress.status === "running"
-                      ? "info"
-                      : progress.status === "completed"
-                      ? "success"
-                      : progress.status === "aborted"
-                      ? "warning"
-                      : "error"
-                  }
-                >
-                  {t(`settings.auto_dream_status_${progress.status}`, progress.status)}
-                </Badge>
-              )}
-            </div>
-            {optedIn ? (
-              <p className="text-[11px] text-text-dim">
-                {t("settings.auto_dream_last", "Last")}:{" "}
-                {formatRelativeMs(agent.last_consolidated_at_ms, now, i18n.language, t)}
-                {" · "}
-                {t("settings.auto_dream_next", "Next")}:{" "}
-                {formatRelativeMs(agent.next_eligible_at_ms, now, i18n.language, t)}
-                {" · "}
-                {agent.effective_min_sessions > 0 ? (
-                  <span
-                    title={t(
-                      "settings.auto_dream_sessions_progress_title",
-                      "Sessions touched since last dream / required threshold",
-                    )}
-                  >
-                    {agent.sessions_since_last}/{agent.effective_min_sessions}{" "}
-                    {t("settings.auto_dream_sessions_since", "sessions since")}
-                  </span>
-                ) : (
-                  <>
-                    {agent.sessions_since_last}{" "}
-                    {t("settings.auto_dream_sessions_since", "sessions since")}
-                  </>
-                )}
-                {" · "}
-                <span
-                  title={t(
-                    "settings.auto_dream_effective_title",
-                    "Resolved threshold — manifest override or global default",
-                  )}
-                >
-                  {t("settings.auto_dream_every", "every")}{" "}
-                  {formatHours(agent.effective_min_hours, t)}
-                </span>
-              </p>
-            ) : running ? (
-              // Agent was toggled off while a manual dream was already in
-              // flight. Keep the operator informed — the run continues to
-              // completion or abort, and the abort button above stays live.
-              <p className="text-[11px] text-text-dim italic">
-                {t(
-                  "settings.auto_dream_opt_out_running",
-                  "Disabled mid-dream — the current run will finish or can be aborted.",
-                )}
-              </p>
-            ) : (
-              <p className="text-[11px] text-text-dim italic">
-                {t(
-                  "settings.auto_dream_opt_in_hint",
-                  "Not enrolled — toggle on to include in the scheduler.",
-                )}
-              </p>
-            )}
-          </div>
-        </div>
-        <div className="flex gap-2 shrink-0 items-center">
-          <label
-            className="flex items-center gap-1.5 cursor-pointer select-none"
-            title={t("settings.auto_dream_toggle_title", "Opt this agent in or out")}
-          >
-            <input
-              type="checkbox"
-              checked={optedIn}
-              disabled={togglePending}
-              onChange={(e) => onToggle(agent.agent_id, e.target.checked)}
-              className="w-3.5 h-3.5 accent-purple-500"
-            />
-            <span className="text-[11px] text-text-dim">
-              {optedIn
-                ? t("settings.auto_dream_enrolled", "Enrolled")
-                : t("settings.auto_dream_not_enrolled", "Off")}
-            </span>
-          </label>
-          {running && agent.can_abort && (
-            // Surface the abort affordance even when the agent has been
-            // toggled off mid-dream — otherwise the in-flight operation
-            // keeps spending tokens with no UI to stop it.
-            <Button
-              variant="secondary"
-              size="sm"
-              onClick={() => onAbort(agent.agent_id)}
-              disabled={abortPending}
-            >
-              <Square className="w-3.5 h-3.5 mr-1.5" />
-              {t("settings.auto_dream_abort", "Abort")}
-            </Button>
-          )}
-          {optedIn && (
-            <Button
-              variant="secondary"
-              size="sm"
-              onClick={() => onTrigger(agent.agent_id)}
-              disabled={triggerPending || disabled || running}
-              title={disabled ? t("settings.auto_dream_off", "Disabled") : undefined}
-            >
-              <Play className="w-3.5 h-3.5 mr-1.5" />
-              {t("settings.auto_dream_trigger", "Dream now")}
-            </Button>
-          )}
-        </div>
-      </div>
-
-      {progress && (progress.status !== "completed" || progress.memories_touched.length > 0) && (
-        <div className="px-3 pb-2 pt-1 border-t border-border-subtle/30 space-y-1">
-          <p className="text-[10px] text-text-dim">
-            <span className="uppercase tracking-wider">
-              {t("settings.auto_dream_phase", "Phase")}:
-            </span>{" "}
-            <span className="font-mono">{progress.phase}</span>
-            {" · "}
-            {progress.tool_use_count}{" "}
-            {t("settings.auto_dream_tool_calls", "tool calls")}
-            {progress.memories_touched.length > 0 && (
-              <>
-                {" · "}
-                {progress.memories_touched.length}{" "}
-                {t("settings.auto_dream_memories_touched", "memories touched")}
-              </>
-            )}
-          </p>
-          {lastTurn && lastTurn.text && (
-            <p className="text-[11px] text-text-muted line-clamp-2 italic">
-              &ldquo;{lastTurn.text}&rdquo;
-            </p>
-          )}
-          {progress.error && (
-            <p className="text-[11px] text-red-500">
-              <XCircle className="w-3 h-3 inline mr-1" />
-              {progress.error}
-            </p>
-          )}
-          {/* Cache-hit visibility. Since the forkedAgent migration, dreams
-              fork off the parent turn and hit Anthropic's prompt cache on
-              the (system + tools + messages) prefix. Surfacing the hit
-              rate here lets operators see the actual cost win — the
-              whole reason the forkedAgent PR exists. Only shown for
-              completed dreams (usage is populated then) and only when
-              there actually was input (avoids 0/0 noise). */}
-          {progress.usage && progress.usage.input_tokens > 0 && (
-            <p className="text-[10px] text-text-dim">
-              <span className="uppercase tracking-wider">
-                {t("settings.auto_dream_cache", "Cache")}:
-              </span>{" "}
-              {(() => {
-                const u = progress.usage!;
-                const totalIn =
-                  u.input_tokens +
-                  u.cache_read_input_tokens +
-                  u.cache_creation_input_tokens;
-                const hitPct =
-                  totalIn > 0
-                    ? Math.round((u.cache_read_input_tokens / totalIn) * 100)
-                    : 0;
-                return (
-                  <span
-                    title={t(
-                      "settings.auto_dream_cache_title",
-                      "Prompt cache hit rate for this dream — higher means more of the prefix came from Anthropic's cache instead of being re-billed.",
-                    )}
-                  >
-                    <span className="font-mono">{hitPct}%</span>
-                    {" "}
-                    ({u.cache_read_input_tokens.toLocaleString()}/
-                    {totalIn.toLocaleString()} tok)
-                  </span>
-                );
-              })()}
-              {typeof progress.usage.cost_usd === "number" && (
-                <>
-                  {" · "}
-                  <span
-                    title={t(
-                      "settings.auto_dream_cost_title",
-                      "Measured provider cost for this dream turn (input + output, cached tokens billed at the reduced rate).",
-                    )}
-                  >
-                    ${progress.usage.cost_usd.toFixed(5)}
-                  </span>
-                </>
-              )}
-            </p>
-          )}
-        </div>
-      )}
-    </div>
-  );
-}
-
-function AutoDreamSection() {
-  const { t } = useTranslation();
-  const statusQuery = useAutoDreamStatus();
-  const trigger = useTriggerAutoDream();
-  const abort = useAbortAutoDream();
-  const setEnabled = useSetAutoDreamEnabled();
-  const [error, setError] = useState<string | null>(null);
-  const [lastMsg, setLastMsg] = useState<string | null>(null);
-
-  const status = statusQuery.data;
-
-  const onTrigger = async (agentId: string) => {
-    setError(null);
-    setLastMsg(null);
-    try {
-      const outcome = await trigger.mutateAsync(agentId);
-      setLastMsg(outcome.fired ? t("settings.auto_dream_fired", "Consolidation fired") : outcome.reason);
-    } catch (e) {
-      setError(e instanceof Error ? e.message : String(e));
-    }
-  };
-
-  const onAbort = async (agentId: string) => {
-    setError(null);
-    setLastMsg(null);
-    try {
-      const outcome = await abort.mutateAsync(agentId);
-      setLastMsg(outcome.aborted ? t("settings.auto_dream_aborted", "Abort signalled") : outcome.reason);
-    } catch (e) {
-      setError(e instanceof Error ? e.message : String(e));
-    }
-  };
-
-  const onToggle = async (agentId: string, enabled: boolean) => {
-    setError(null);
-    setLastMsg(null);
-    try {
-      await setEnabled.mutateAsync({ agentId, enabled });
-      setLastMsg(
-        enabled
-          ? t("settings.auto_dream_enrolled_ok", "Agent enrolled")
-          : t("settings.auto_dream_unenrolled_ok", "Agent unenrolled"),
-      );
-    } catch (e) {
-      setError(e instanceof Error ? e.message : String(e));
-    }
-  };
-
-  return (
-    <div className="rounded-2xl border border-border-subtle bg-surface">
-      <div className="px-5 py-3 border-b border-border-subtle/50 flex items-center justify-between">
-        <p className="text-[10px] font-black uppercase tracking-widest text-text-dim">
-          {t("settings.auto_dream", "Auto-Dream")}
-        </p>
-        {status && (
-          <Badge variant={status.enabled ? "success" : "default"}>
-            {status.enabled
-              ? t("settings.auto_dream_on", "Enabled")
-              : t("settings.auto_dream_off", "Disabled")}
-          </Badge>
-        )}
-      </div>
-      <div className="px-5 py-3">
-        <p className="text-xs text-text-dim mb-3">
-          {t(
-            "settings.auto_dream_desc",
-            "Periodically asks opt-in agents to consolidate their memory. Configure in config.toml via [auto_dream] enabled + per-agent auto_dream_enabled.",
-          )}
-        </p>
-
-        {statusQuery.isLoading && (
-          <p className="text-xs text-text-dim">{t("common.loading", "Loading…")}</p>
-        )}
-        {statusQuery.isError && (
-          <p className="text-xs text-red-500">
-            {t("settings.auto_dream_load_err", "Failed to load auto-dream status")}
-          </p>
-        )}
-
-        {status && status.agents.length === 0 && (
-          <p className="text-xs text-text-dim italic">
-            {t(
-              "settings.auto_dream_no_agents_registered",
-              "No agents registered yet. Create an agent first, then toggle it on here.",
-            )}
-          </p>
-        )}
-
-        {status && status.agents.length > 0 && (
-          <div className="space-y-2">
-            {status.agents.map((a) => (
-              <AutoDreamAgentRow
-                key={a.agent_id}
-                agent={a}
-                disabled={!status.enabled}
-                onTrigger={onTrigger}
-                onAbort={onAbort}
-                onToggle={onToggle}
-                triggerPending={trigger.isPending && trigger.variables === a.agent_id}
-                abortPending={abort.isPending && abort.variables === a.agent_id}
-                togglePending={setEnabled.isPending && setEnabled.variables?.agentId === a.agent_id}
-              />
-            ))}
-          </div>
-        )}
-
-        {lastMsg && (
-          <p className="text-xs text-green-500 mt-2">
-            <CheckCircle className="w-3 h-3 inline mr-1" />
-            {lastMsg}
-          </p>
-        )}
-        {error && (
-          <p className="text-xs text-red-500 mt-2">
-            <XCircle className="w-3 h-3 inline mr-1" />
-            {error}
-          </p>
-        )}
       </div>
     </div>
   );
